carry-agent/requirements.txt
alabaster==0.7.12
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.3.3
asttokens==1.1.13
automationhat==0.2.0
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.7.1
blinker==1.4
blinkt==0.1.2
bluepy==1.3.0
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2018.8.24
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.3.7
colorzero==1.1
commonmark==0.9.1
cookies==2.2.1
cryptography==2.6.1
cycler==0.10.0
decorator==4.3.0
distlib==0.3.0
dlib==19.19.0
dnspython==1.16.0
docutils==0.14
drumhat==0.1.0
email-validator==1.1.0
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==1.0.2
Flask-Admin==1.5.6
Flask-Bcrypt==0.7.1
flask-googlemaps==0.4.0
Flask-Login==0.5.0
flask-marshmallow==0.12.0
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
fourletterphat==0.1.0
gevent==20.5.0
gevent-websocket==0.10.1
googlemaps==4.3.1
gpiozero==1.5.1
greenlet==0.4.15
guizero==0.6.0
html5lib==1.0.1
idna==2.6
imagesize==1.2.0
importlib-metadata==1.6.0
imutils==0.5.3
ipykernel==4.9.0
ipython==5.8.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
jupyter-client==5.2.3
jupyter-core==4.4.0
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.0.1
lazy-object-proxy==1.4.3
logilab-common==1.4.2
lxml==4.3.2
MarkupSafe==1.1.0
marshmallow==3.6.0
marshmallow-sqlalchemy==0.23.0
matplotlib==3.0.2
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mypy==0.670
mypy-extensions==0.4.1
mysqlclient==1.4.6
nudatus==0.0.4
numpy==1.16.2
oauthlib==2.1.0
olefile==0.46
opencv-contrib-python==4.1.0.25
packaging==20.4
pantilthat==0.0.7
parso==0.3.1
pbr==5.4.5
pexpect==4.6.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
pickleshare==0.7.5
picraft==1.0
piglow==1.2.5
pigpio==1.44
Pillow==5.4.1
prompt-toolkit==1.0.15
psutil==5.5.1
PyAudio==0.2.11
PyBluez==0.23
pycairo==1.16.2
pycodestyle==2.4.0
pycparser==2.20
pycrypto==2.6.1
pyflakes==2.0.0
pygame==1.9.4.post1
Pygments==2.6.1
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.4.4
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.2.0
pypng==0.0.20
PyQRCode==1.2.1
pyserial==3.4
python-apt==1.8.4.1
python-crontab==2.4.1
python-dateutil==2.7.3
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2020.1
pyxdg==0.25
pyzbar==0.1.8
pyzmq==17.1.2
qtconsole==4.3.1
rainbowhat==0.1.0
recommonmark==0.6.0
requests==2.21.0
requests-oauthlib==1.0.0
requirements==0.1
responses==0.9.0
rinoh-typeface-dejavuserif==0.1.1
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.4.0
roman==2.0.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scipy==1.4.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
semver==2.0.1
Send2Trash==1.5.0
sense-emu==1.1
sense-hat==2.2.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
skywriter==0.0.7
sn3218==1.2.7
snowballstemmer==2.0.0
soupsieve==1.8
SpeechRecognition==3.8.1
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spidev==3.4
SQLAlchemy==1.3.16
sqlalchemy-filters==0.11.0
ssh-import-id==5.7
stevedore==1.32.0
thonny==3.2.6
tornado==5.1.1
touchphat==0.0.1
traitlets==4.3.2
twython==3.7.0
typed-ast==1.4.1
uflash==1.2.4
unicornhathd==0.0.4
urllib3==1.24.1
virtualenv==20.0.18
virtualenv-clone==0.5.4
virtualenvwrapper==5.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.11.2
WTForms==2.3.1
zipp==3.1.0
